+++ shuup/admin/modules/sales_dashboard/__init__.py
@@ -27,3 +27,4 @@ class SalesDashboardModule(CurrencyBound, AdminModule):
         yield dashboard.get_avg_purchase_size_block(request, currency)
         yield dashboard.get_open_orders_block(request, currency)
         yield dashboard.get_order_value_chart_dashboard_block(request, currency)
+        yield dashboard.get_shop_overview_block(request, currency)